requirements.txt
Django==1.7.1
django-grappelli
MySQL-python
djangorestframework
